Axis to Button functionality (recording and assigning) should work now. Settings are not saved online yet. Default deadzone range increased to [0;32767] in order to match with DirectInput precision. Default deadzone is set to 25% (8192).

+++ x360ce/x360ce/Config.h
@@ -56,7 +56,7 @@ public:
 		MappingType type;
 		s8 id;			// Index for the mapped button/axis/slider
 		s8 but;
-		u8 buttondz;
+		u16 buttondz;
 		ButtonMap()
 		{
 			id = 0;
